# Conflicts:
#	260-delegate/src/main/java/software/wings/delegatetasks/azure/arm/deployment/validator/Validators.java
#	260-delegate/src/main/java/software/wings/delegatetasks/azure/common/validator/Validator.java
#	950-delegate-tasks-beans/src/main/java/io/harness/delegate/beans/azure/appservicesettings/value/AzureAppServiceHarnessSettingSecretRef.java

+++ 960-ng-core-beans/src/main/java/io/harness/ng/core/dto/ActiveProjectsCountDTO.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright 2021 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Shield 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
+ */
+
 package io.harness.ng.core.dto;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
@@ -16,5 +23,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Schema(name = "ActiveProjectsCount", description = "This gives the count of the active Projects")
 public class ActiveProjectsCountDTO {
-  Integer count;
+  long count;
 }
